Prettify programming code
